move prop-types to the modules/references/client
