+++ modules/references/client/components/read-references/Reference.js
@@ -11,7 +11,7 @@ import Avatar from '@/modules/users/client/components/Avatar.component';
 import Meta from './Meta';
 import TimeAgo from '@/modules/core/client/components/TimeAgo';
 import UserLink from '@/modules/users/client/components/UserLink';
-import { experienceType } from '../../../experiences.prop-types';
+import { experienceType } from '../../experiences.prop-types';
 
 const PendingNotice = styled.div`
   font-style: italic;
